Separated Sentry into its own install command
